requirements.txt
pydantic>=2.7.0
python-dotenv>=1.0.0
pypdf>=4.0.0

sentence-transformers>=3.0.0 # free local embedding model (SBERT)

tiktoken>=0.7.0

chromadb>=0.5.0

openai>=1.40.0

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
